refactor(editor): split tiptap.ts god-file by concern
Slice #4 of docs/tiptap-rebuild-spec.md. Pure file moves, no behavior change.

The 1267-line tiptap.ts mixed six concerns. It is now split under the existing
src/lib/tiptap/ directory:

  index.ts          barrel — the public import surface
  content.ts        trimContent, isContentEmpty, excalidraw-chip walkers
  serialize.ts      extractContentNodes, contentNodesToTipTap, chip registry
  suggestions.ts    createSuggestionRender + the tag/slash suggestion configs
  replace-fence.ts  parseFenceFromJson + the fence<->preview transforms
  extensions/EditorShortcuts.ts
  extensions/AnnotBulletList.ts
  extensions/PasteHandlers.ts   (ImagePasteHandler, TextPasteHandler, shouldChip)
  extensions/SlashCommands.ts

The ./tiptap import path now resolves to tiptap/index.ts, so all five
importers are unchanged. Module-private helpers (MARK_WRAPPERS,
TextAccumulator, etc.) stay unexported.

Note: replaceExcalidrawChip and createTagSuggestion moved verbatim but have no
callers — dead-export cleanup left for a later pass.
