+++ src/lib/tiptap/extensions/EditorShortcuts.ts
@@ -1,0 +1,34 @@
+import { Extension } from '@tiptap/core';
+
+/**
+ * EditorShortcuts extension - handles keyboard shortcuts at the TipTap level
+ * to prevent default behavior from firing first.
+ */
+export interface EditorShortcutsOptions {
+  onSubmit?: () => void;
+  onDismiss?: () => void;
+}
+
+export const EditorShortcuts = Extension.create<EditorShortcutsOptions>({
+  name: 'editorShortcuts',
+
+  addOptions() {
+    return {
+      onSubmit: undefined,
+      onDismiss: undefined,
+    };
+  },
+
+  addKeyboardShortcuts() {
+    return {
+      'Mod-Enter': () => {
+        this.options.onSubmit?.();
+        return true; // Prevent default Enter behavior
+      },
+      Escape: () => {
+        this.options.onDismiss?.();
+        return true;
+      },
+    };
+  },
+});
