+++ src/lib/tiptap/extensions/SlashCommands.ts
@@ -1,0 +1,50 @@
+import { Extension } from '@tiptap/core';
+import { PluginKey } from '@tiptap/pm/state';
+import Suggestion, { type SuggestionOptions } from '@tiptap/suggestion';
+
+// Unique plugin key for slash command suggestions
+const SlashSuggestionPluginKey = new PluginKey('slashSuggestion');
+
+/**
+ * SlashCommand interface for extensible slash commands.
+ */
+export interface SlashCommand {
+  id: string;
+  name: string;
+  description: string;
+  icon: string;
+  action: (editor: import('@tiptap/core').Editor, range: import('@tiptap/core').Range) => void;
+}
+
+/**
+ * SlashCommands extension - provides `/` triggered command menu.
+ */
+export interface SlashCommandsOptions {
+  suggestion: Omit<SuggestionOptions<SlashCommand>, 'editor' | 'pluginKey'>;
+}
+
+export const SlashCommands = Extension.create<SlashCommandsOptions>({
+  name: 'slashCommands',
+
+  addOptions() {
+    return {
+      suggestion: {
+        char: '/',
+        items: () => [],
+        command: ({ editor, range, props }) => {
+          props.action(editor, range);
+        },
+      },
+    };
+  },
+
+  addProseMirrorPlugins() {
+    return [
+      Suggestion({
+        editor: this.editor,
+        pluginKey: SlashSuggestionPluginKey,
+        ...this.options.suggestion,
+      }),
+    ];
+  },
+});
